+++ src/warp.h
@@ -4,7 +4,7 @@
 		float4	position	: POSITION;
 		float4  inscatter	: TEXCOORD0;
 		float4	extinction	: TEXCOORD1;
-		half4	texcoord12	: TEXCOORD2;
+		float4	texcoord12	: TEXCOORD2;
 		float2  params		: TEXCOORD3;
 	};
      
@@ -12,6 +12,6 @@
 		float4	position	: POSITION;
 		float3	inscatter	: TEXCOORD0;
 		float3	extinction	: TEXCOORD1;
-		half3	eye			: TEXCOORD2;
-		half2	texCoord	: TEXCOORD3;
+		float3	eye			: TEXCOORD2;
+		float2	texCoord	: TEXCOORD3;
 	};
